# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: B2B AI Marketing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /licenses:
    get:
      operationId: getLicenses
      tags: [licenses]
      summary: List all licenses / license pools
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [seed, purchase, null]
      responses:
        "200":
          description: License list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/License"
    post:
      operationId: createLicense
      tags: [licenses]
      summary: Create a new license pool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseInput"
      responses:
        "201":
          description: Created license
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /licenses/{id}:
    get:
      operationId: getLicense
      tags: [licenses]
      summary: Get a license pool by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: License
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "404":
          description: Not found
    patch:
      operationId: updateLicense
      tags: [licenses]
      summary: Update a license pool (threshold, settings)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseUpdate"
      responses:
        "200":
          description: Updated license
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
    delete:
      operationId: deleteLicense
      tags: [licenses]
      summary: Delete a license pool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /licenses/pool-summary:
    get:
      operationId: getLicensePoolSummary
      tags: [licenses]
      summary: Get aggregated pool summary (active, inactive, available, at-risk)
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Pool summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicensePoolSummary"

  # ── License Seats ─────────────────────────────────────────────────────────────
  /seats:
    get:
      operationId: getSeats
      tags: [licenses]
      summary: List license seat assignments
      parameters:
        - name: licenseId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [active, inactive, reserved, at_risk, null]
      responses:
        "200":
          description: Seat list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Seat"
  /seats/{id}/assign:
    post:
      operationId: assignSeat
      tags: [licenses]
      summary: Assign a seat to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatAssignInput"
      responses:
        "200":
          description: Updated seat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seat"
  /seats/{id}/release:
    post:
      operationId: releaseSeat
      tags: [licenses]
      summary: Manually release a seat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Released seat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seat"
  /seats/{id}/revoke:
    post:
      operationId: revokeSeat
      tags: [licenses]
      summary: Revoke a seat (admin override)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revoked seat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seat"

  # ── Audit Logs ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    License:
      type: object
      required: [id, name, type, clientId, totalSeats, activeSeats, inactiveSeats, availableSeats, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [seed, purchase]
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        totalSeats:
          type: integer
        activeSeats:
          type: integer
        inactiveSeats:
          type: integer
        availableSeats:
          type: integer
        reservedSeats:
          type: integer
        atRiskSeats:
          type: integer
        status:
          type: string
          enum: [healthy, at_risk, over_capacity, inactive]
        inactivityThresholdDays:
          type: ["integer", "null"]
        autoRelease:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    LicenseInput:
      type: object
      required: [name, type, clientId, totalSeats]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [seed, purchase]
        clientId:
          type: integer
        totalSeats:
          type: integer
        inactivityThresholdDays:
          type: ["integer", "null"]
        autoRelease:
          type: boolean
        expiresAt:
          type: ["string", "null"]
    LicenseUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        totalSeats:
          type: ["integer", "null"]
        inactivityThresholdDays:
          type: ["integer", "null"]
        autoRelease:
          type: ["boolean", "null"]
        status:
          type: ["string", "null"]
          enum: [healthy, at_risk, over_capacity, inactive, null]
    LicensePoolSummary:
      type: object
      required: [totalLicenses, totalSeats, activeSeats, inactiveSeats, availableSeats, atRiskSeats, utilizationRate]
      properties:
        totalLicenses:
          type: integer
        totalSeats:
          type: integer
        activeSeats:
          type: integer
        inactiveSeats:
          type: integer
        availableSeats:
          type: integer
        atRiskSeats:
          type: integer
        utilizationRate:
          type: number
        seedSeats:
          type: integer
        purchaseSeats:
          type: integer

    # ── Seat Schemas ───────────────────────────────────────────────────────────
    Seat:
      type: object
      required: [id, licenseId, status, createdAt]
      properties:
        id:
          type: integer
        licenseId:
          type: integer
        licenseName:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, reserved, at_risk, available]
        assignedAt:
          type: ["string", "null"]
        lastActiveAt:
          type: ["string", "null"]
        inactiveDays:
          type: ["integer", "null"]
        createdAt:
          type: string
    SeatAssignInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer

    # ── Audit Log Schemas ──────────────────────────────────────────────────────
...