# lib/api-zod/src/generated/types/licenseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * B2B AI Marketing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LicenseInputType } from './licenseInputType';

export interface LicenseInput {
  name: string;
  type: LicenseInputType;
  clientId: number;
  totalSeats: number;
  /** @nullable */
  inactivityThresholdDays?: number | null;
  autoRelease?: boolean;
  /** @nullable */
  expiresAt?: string | null;
}