# lib/api-zod/src/generated/types/seat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * B2B AI Marketing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SeatStatus } from './seatStatus';

export interface Seat {
  id: number;
  licenseId: number;
  /** @nullable */
  licenseName?: string | null;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  status: SeatStatus;
  /** @nullable */
  assignedAt?: string | null;
  /** @nullable */
  lastActiveAt?: string | null;
  /** @nullable */
  inactiveDays?: number | null;
  createdAt: string;
}